Renaming of from Plugin -> EffectPlugin before creating GenericPlugin Also add ctrl-c handler
