+++ turbo_audio/Cargo.toml
@@ -11,6 +11,7 @@ bytemuck = { version = "1.14.0", features = ["derive"] }
 chrono = "0.4.19"
 clap = { version = "4.4.8", features = ["derive"] }
 cpal = { version = "0.15.2" }
+ctrlc = "3.4.4"
 dasp = "0.11.0"
 dasp_ring_buffer = "0.11.0"
 dasp_signal = "0.11.0"
